homeページとLoginページ、registerページの編集
